+++ src/settingsdialog/typehandler.cpp
@@ -24,7 +24,7 @@ TypeHandler::~TypeHandler ()
 QWidget* TypeHandler::operator() (const QVariant& var, const SettingsItemInfo& sii, const QString& propName, QObject* owner)
 {
     for (int i = 0; i < Converters_.size (); ++i)
-        if (var.userType () == Converters_ [i]->GetType ())
+        if (var.userType () == static_cast<int> (Converters_ [i]->GetType ()))
             return Converters_ [i]->Convert (var, sii, propName, owner);
     return 0;
 }
